+++ draft-packages/form/KaizenDraft/Form/Primitives/Checkbox/Checkbox.tsx
@@ -4,7 +4,7 @@ import { OverrideClassName } from "@kaizen/component-base"
 import { Icon } from "@kaizen/component-library"
 import checkIcon from "@kaizen/component-library/icons/check.icon.svg"
 import minusIcon from "@kaizen/component-library/icons/minus.icon.svg"
-import styles from "./styles.scss"
+import styles from "./Checkbox.module.scss"
 
 export type CheckedStatus = "on" | "off" | "mixed"