+++ draft-packages/likert-scale-legacy/KaizenDraft/LikertScaleLegacy/LikertScaleLegacy.tsx
@@ -1,11 +1,9 @@
 import React, { useState, createRef } from "react"
-
 import classnames from "classnames"
 import { Paragraph } from "@kaizen/typography"
 import determineSelectionFromKeyPress from "./helpers/determineSelectionFromKeyPress"
 import { Scale, ScaleItem, ScaleValue } from "./types"
-
-import styles from "./styles.module.scss"
+import styles from "./LikertScaleLegacy.module.scss"
 
 type ItemRefs = Array<{
   value: ScaleValue