+++ draft-packages/filter-menu-button/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.5.20](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-filter-menu-button@2.5.19...@kaizen/draft-filter-menu-button@2.5.20) (2022-07-27)
+
+**Note:** Version bump only for package @kaizen/draft-filter-menu-button
+
+
+
+
+
+## [2.5.19](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-filter-menu-button@2.5.18...@kaizen/draft-filter-menu-button@2.5.19) (2022-07-27)
+
+**Note:** Version bump only for package @kaizen/draft-filter-menu-button
+
+
+
+
+
 ## [2.5.18](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-filter-menu-button@2.5.17...@kaizen/draft-filter-menu-button@2.5.18) (2022-07-26)
 
 **Note:** Version bump only for package @kaizen/draft-filter-menu-button